+++ src/Piral.Blazor.Orchestrator/Connector/MfEmptyConnector.cs
@@ -5,7 +5,7 @@ namespace Piral.Blazor.Orchestrator.Connector;
 
 internal class MfEmptyConnector : IMfDebugConnector
 {
-    private readonly IEnumerable<string> empty = Enumerable.Empty<string>();
+    private readonly IEnumerable<string> empty = [];
     private readonly Task<bool> notIntercepted = Task.FromResult(false);
 
     public IEnumerable<string> Styles => empty;