+++ src/Piral.Blazor.Orchestrator/Loader/MfLocalLoaderService.cs
@@ -2,21 +2,13 @@
 
 namespace Piral.Blazor.Orchestrator.Loader;
 
-internal class MfLocalLoaderService<T> : IMfLoaderService
+internal class MfLocalLoaderService<T>(T originalLoader, IMfRepository repository, IModuleContainerService container, IEvents events) : IMfLoaderService
     where T : class, IMfLoaderService
 {
-    private readonly T _originalLoader;
-    private readonly IMfRepository _repository;
-    private readonly IModuleContainerService _container;
-    private readonly IEvents _events;
-
-    public MfLocalLoaderService(T originalLoader, IMfRepository repository, IModuleContainerService container, IEvents events)
-    {
-        _originalLoader = originalLoader;
-        _repository = repository;
-        _container = container;
-        _events = events;
-    }
+    private readonly T _originalLoader = originalLoader;
+    private readonly IMfRepository _repository = repository;
+    private readonly IModuleContainerService _container = container;
+    private readonly IEvents _events = events;
 
     public void ConnectMicrofrontends(CancellationToken cancellationToken)
     {